requirements.txt
requests
psycopg2-binary # Use binary for easier installation without system compilers